requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.1
pygithub==2.1.1
playwright==1.40.0
python-multipart==0.0.6
